refactor: refactor code for sprint-2
