+++ app/routes/notFoundPage/index.tsx
@@ -5,7 +5,7 @@ import { AppRoutes } from '@/sources/enums';
 import { notFoundPage as messages } from '@/sources/messages/notFoundPage';
 
 import type { Route } from '../../+types/root';
-import styles from './notFoundPage.module.css';
+import styles from './NotFoundPage.module.css';
 
 export function meta({ location }: Route.MetaArgs) {
   return [
