+++ app/routes/privateRoutes/historyPage/index.tsx
@@ -4,7 +4,7 @@ import { historyPageMessages as messages } from '@/sources/messages/historyPage.
 
 import { pageMeta } from '@/utils/metaHelpers.ts';
 
-import styles from './history.module.css';
+import styles from './History.module.css';
 
 export const meta = pageMeta(messages);
