+++ app/routes/privateRoutes/restClientPage/components/codeGenerator/index.tsx
@@ -1,16 +1,20 @@
-import { selectCode } from '@/store/slices/restClient/selectors';
+import {
+  selectCode,
+  selectLanguage,
+} from '@/store/slices/restClient/selectors';
 
 import type { FC } from 'react';
 
 import { useSelector } from 'react-redux';
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import vscDarkPlus from 'react-syntax-highlighter/dist/esm/styles/prism/vsc-dark-plus';
+import { atomDark } from 'react-syntax-highlighter/dist/cjs/styles/prism';
 
 export const CodeGenerator: FC = () => {
   const code = useSelector(selectCode);
+  const language = useSelector(selectLanguage);
 
   return (
-    <SyntaxHighlighter language="javascript" style={vscDarkPlus}>
+    <SyntaxHighlighter language={language?.toLowerCase()} style={atomDark}>
       {code?.generatedCode || ''}
     </SyntaxHighlighter>
   );
