+++ app/routes/privateRoutes/restClientPage/components/headers/index.tsx
@@ -8,22 +8,18 @@ import { KeyValueEditor } from '@/components/keyValueEditor';
 
 import { useActions } from '@/utils/hooks/useActions';
 
-import styles from './Headers.module.css';
-
 export const Headers = () => {
   const headers = useSelector(selectHeaders);
   const { addHeader, updateHeader, removeHeader } = useActions();
 
   return (
-    <div className={styles.container}>
-      <KeyValueEditor
-        keyHeader={messages.table.headerKey}
-        valueHeader={messages.table.headerValue}
-        keyValues={headers}
-        onAdd={addHeader}
-        onDelete={removeHeader}
-        onUpdate={updateHeader}
-      />
-    </div>
+    <KeyValueEditor
+      keyHeader={messages.table.headerKey}
+      valueHeader={messages.table.headerValue}
+      keyValues={headers}
+      onAdd={addHeader}
+      onDelete={removeHeader}
+      onUpdate={updateHeader}
+    />
   );
 };
