+++ app/routes/privateRoutes/variablesPage/index.tsx
@@ -21,16 +21,15 @@ export default function VariablesPage() {
     <div className={styles.wrapper}>
       <div className={styles.container}>
         <h1 className={styles.header}>{messages.header}</h1>
-        <div>
-          <KeyValueEditor
-            keyHeader={messages.keyHeader}
-            valueHeader={messages.valueHeader}
-            keyValues={variables}
-            onAdd={addVariable}
-            onUpdate={updateVariable}
-            onDelete={removeVariable}
-          />
-        </div>
+
+        <KeyValueEditor
+          keyHeader={messages.keyHeader}
+          valueHeader={messages.valueHeader}
+          keyValues={variables}
+          onAdd={addVariable}
+          onUpdate={updateVariable}
+          onDelete={removeVariable}
+        />
       </div>
     </div>
   );
