--- app/types/postman-code-generators.d.ts
@@ -1,1 +1,0 @@
-declare module 'postman-code-generators';
